# api/openapi.yaml
openapi: "3.0.3"
info:
  title: ACE-Step Suno-Compatible API
  version: "1.5.0"
  description: |
    100% Suno-API-compatible REST interface powered by ACE-Step 1.5.

    All music-generation endpoints are **asynchronous**: the initial POST returns a
    `taskId` that you poll with the corresponding `record-info` GET until `status`
    reaches `SUCCESS` or `FAILED`.

    Output audio files are served under `/audio/<filename>` as static files.

    ### Authentication
    When the server is configured with `API_KEY`, every request **must** include:
    ```
    Authorization: Bearer <api_key>
    ```
    If `API_KEY` is unset, the header is optional and all requests are accepted.

  contact:
    name: ACE-Step Project
    url: https://github.com/ace-step/ACE-Step-1.5
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:7080
    description: Local development server

# ─────────────────────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: >
        Pass your API key as a Bearer token.
        Leave header absent when the server has no API_KEY configured.

  # ── Shared enums ────────────────────────────────────────────────────────────
  schemas:
...
    SeparationType:
      type: string
      enum: [separate_vocal, split_stem]
      default: separate_vocal
      description: |
        - `separate_vocal`: isolate the main vocal track
        - `split_stem`: split into multiple stems (drums, bass, melody, vocals)
...
    TaskStatus:
      type: string
      description: |
        Lifecycle of an async task:
        - `PENDING` – queued, not yet started
        - `GENERATING` – actively processing
        - `TEXT_SUCCESS` – LLM phase completed, DiT phase pending
        - `FIRST_SUCCESS` – first audio sample ready (batch mode)
        - `SUCCESS` – all outputs ready
        - `FAILED` – terminal error
        - `CREATE_TASK_FAILED` – task could not be enqueued
      enum:
        - PENDING
        - GENERATING
        - TEXT_SUCCESS
        - FIRST_SUCCESS
        - SUCCESS
        - FAILED
        - CREATE_TASK_FAILED
    TaskType:
      type: string
      enum: [GENERATE, LYRICS, VOCAL_REMOVAL, WAV, MP4, COVER, SOUNDS]

    # ── Common response envelope ───────────────────────────────────────────────
    ApiResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          description: HTTP-like status code (200 = success).
          example: 200
        msg:
          type: string
          example: success
        data:
          description: Payload; null on error.
          nullable: true
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            code:
              example: 400
            msg:
              example: Bad Request
            data:
              nullable: true
              example: null
    TaskCreatedData:
      type: object
      required: [taskId]
      properties:
        taskId:
          type: string
          description: Opaque identifier – use it with `record-info` endpoints.
          example: a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c

    # ── Audio output entry (sunoData item) ────────────────────────────────────
    SunoAudioEntry:
      type: object
      properties:
        id:
          type: string
          description: Unique audio identifier (UUID hex).
          example: d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9
        audioUrl:
          type: string
          nullable: true
          description: Relative URL to stream / download the audio file.
          example: /audio/a3f1b24c_0.flac
        streamAudioUrl:
          type: string
          nullable: true
          example: /audio/a3f1b24c_0.flac
        imageUrl:
          type: string
          nullable: true
          description: Cover art URL (null unless cover was generated).
        prompt:
          type: string
          description: Lyrics or caption embedded in the output.
        modelName:
          type: string
          example: ACE-Step-v1.5
        title:
          type: string
          example: Track 1
        tags:
          type: string
          description: Style tags used for generation.
        createTime:
          type: string
          description: UTC timestamp of generation.
          example: "2026-04-11 12:00:00"
        duration:
          type: number
          nullable: true
          description: Duration of the audio in seconds.
          example: 30.4

    # ── Task status response ───────────────────────────────────────────────────
    TaskDetailData:
      type: object
      required: [taskId, status, type]
      properties:
        taskId:
          type: string
        param:
          type: string
          description: JSON-serialised original request parameters.
        response:
          nullable: true
          description: Generation result payload (sunoData array) once SUCCESS.
          type: object
          properties:
            taskId:
              type: string
            sunoData:
              type: array
              items:
                $ref: "#/components/schemas/SunoAudioEntry"
        status:
          $ref: "#/components/schemas/TaskStatus"
        type:
          $ref: "#/components/schemas/TaskType"
        errorCode:
          type: integer
          nullable: true
        errorMessage:
          type: string
          nullable: true

    # ── Request bodies ─────────────────────────────────────────────────────────
...
    GenerateLyricsRequest:
      type: object
      required: [prompt, callBackUrl]
      properties:
        prompt:
          type: string
          maxLength: 200
          description: Theme or mood description for lyric generation.
          example: "A melancholic song about rainy autumn evenings"
        callBackUrl:
          type: string
          format: uri
    SeparateVocalsRequest:
      type: object
      required: [taskId, audioId]
      properties:
        taskId:
          type: string
          description: Task ID of the completed generation job.
        audioId:
          type: string
          description: Audio entry ID of the track to separate.
        callBackUrl:
          type: string
          format: uri
          nullable: true
        type:
          $ref: "#/components/schemas/SeparationType"
...
    CreateMusicVideoRequest:
      type: object
      required: [taskId, audioId]
      properties:
        taskId:
          type: string
        audioId:
          type: string
        callBackUrl:
          type: string
          format: uri
          nullable: true
        title:
          type: string
          maxLength: 100
          nullable: true
        author:
          type: string
          maxLength: 50
          nullable: true
        domainName:
          type: string
          maxLength: 50
          nullable: true
...
    BoostStyleRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: >
            Short style description to expand, e.g. `"Pop, Mysterious"`.
            Accessible as `style` internally.
          example: "Dark Synthwave, Cinematic, Tense"
        callBackUrl:
          type: string
          format: uri
          nullable: true

    # ── Specialised response payloads ─────────────────────────────────────────
...
    LyricsEntryData:
      type: object
      properties:
        text:
          type: string
          description: Generated lyrics text.
        title:
          type: string
        status:
          type: string
          example: complete
        errorMessage:
          type: string
...
  responses:
    TaskCreated:
      description: Task accepted; poll `record-info` with the returned `taskId`.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TaskCreatedData"
          example:
            code: 200
            msg: success
            data:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
    TaskDetail:
      description: Current task status and output (if complete).
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TaskDetailData"
    BadRequest:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 400
            msg: "customMode=true requires: style, title"
            data: null
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            msg: "Invalid API key."
            data: null
    NotFound:
      description: Task or resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 404
            msg: "Task not found"
            data: null
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/lyrics:
    post:
      summary: Generate lyrics
      operationId: generate_lyrics
      tags: [Lyrics]
      description: |
        Uses the ACE-Step 5Hz LM to generate lyrics from a theme / mood prompt.
        The LLM runs `infer_type="dit"` (metadata-only, no audio codes)
        so this is significantly faster than full music generation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateLyricsRequest"
            example:
              prompt: "A bittersweet farewell to summer"
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/lyrics/record-info:
    get:
      summary: Poll lyrics task
      operationId: get_lyrics_details
      tags: [Lyrics]
      parameters:
        - name: taskId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lyrics task status.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/TaskDetailData"
                          - type: object
                            properties:
                              response:
                                type: object
                                properties:
                                  taskId:
                                    type: string
                                  data:
                                    type: array
                                    items:
                                      $ref: "#/components/schemas/LyricsEntryData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Vocal Removal ───────────────────────────────────────────────────────────
  /api/v1/vocal-removal/generate:
    post:
      summary: Separate vocals / split stems
      operationId: separate_vocals
      tags: [Vocal Removal]
      description: |
        Runs ACE-Step `extract` task on a previously generated track.
        The `audioId` must reference a track from a completed `GENERATE` task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeparateVocalsRequest"
            example:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
              audioId: "d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
              type: separate_vocal
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/vocal-removal/record-info:
    get:
      summary: Poll vocal removal task
      operationId: get_vocal_removal_details
      tags: [Vocal Removal]
      parameters:
        - name: taskId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/TaskDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── WAV ─────────────────────────────────────────────────────────────────────
...
  /api/v1/mp4/generate:
    post:
      summary: Generate music video (MP4)
      operationId: create_music_video
      tags: [Video]
      description: |
        Renders a waveform-visualisation MP4 from a previously generated audio track.
        Uses ffmpeg with `showwaves` filter + optional title overlay.
        Falls back to audio-only MP4 if the waveform filter fails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMusicVideoRequest"
            example:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
              audioId: "d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
              title: "Light Ahead"
              author: "ACE-Step"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/mp4/record-info:
    get:
      summary: Poll MP4 generation task
      operationId: get_mp4_details
      tags: [Video]
      parameters:
        - name: taskId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/TaskDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Cover Art ───────────────────────────────────────────────────────────────
...
  /api/v1/style/generate:
    post:
      summary: Boost style description
      operationId: boost_style
      tags: [Style]
      description: |
        Expands a short style tag string into a richer music production brief
        using the ACE-Step 5Hz LM (`infer_type="dit"`, CoT caption mode).
        Returns the original and expanded versions in the task response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoostStyleRequest"
            example:
              content: "Dark Synthwave, Cinematic, Tense"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/midi/record-info:
    get:
      summary: Poll MIDI task (stub)
      operationId: get_midi_details
      tags: [Style]
      description: |
        MIDI conversion is not currently implemented.
        Returns the stored task state for the given `taskId`.
      parameters:
        - name: taskId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/TaskDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Static audio files ───────────────────────────────────────────────────────
...